+++ data-raw/hab_satellite.R
@@ -1,8 +1,9 @@
 # Code to prepare HAB satellite data sets:
-  # 1) `hab_sat_fr_mil` - counts of pixel values within each CI category for Franks Tract
-    # and Mildred Island
-  # 2) Kept placeholder code to perform same procedure for the EDB regions, but no data object
-    # exists for this currently
+# 1) `hab_sat_ow_delta` - Pixel counts within each Cyano Index (CI) category for
+  # 4 open water regions in the upper San Francisco Estuary: Franks Tract, Mildred
+  # Island, Clifton Court Forebay, and Liberty Island. Also includes a calculated
+  # average Cyano Index value for each region and date in the data set. Summary
+  # statistics cover the spring through late fall (May-Dec) in 2020 and 2021.
 
 # Load packages
 library(dplyr)
@@ -35,7 +36,7 @@ if (download) {
   # Define subfolder directory to store .tif files
   dir_hab_sat <- here("data-raw/HAB_satellite_data")
 
-  # Function to download and unzip harmful algal bloom (HAB) satellite data (cyanobacteria abundance)
+  # Function to download and unzip harmful algal bloom (HAB) satellite data (Cyano Index)
     # from the https://fhab.sfei.org/ website
   download_hab <- function(hab_yr, hab_month) {
     hab_url <- glue("https://fhab.sfei.org/lib/download.php?request=download&dltype=month&year={hab_yr}&month={hab_month}&product=Mosaic")
@@ -76,28 +77,20 @@ df_hab_sat <-
   ) %>%
   select(strs_date, strs_prx_obj)
 
-# Import the polygon shapefiles for Franks-Mildred and the EDB regions
-sf_franks_mildred <- read_sf(here("data-raw/Spatial_data/Franks_Mildred.shp"))
-sf_edb_reg <- read_sf(here("data-raw/Spatial_data/EDB_Regions.shp"))
+# Import the polygon shapefile for the four open water regions in the Delta
+sf_ow_delta <- read_sf(here("data-raw/Spatial_data/Franks_Mildr_CCF_LibIsl.shp"))
 
-# Clean up Franks-Mildred and EDB regions shapefiles to contain just the necessary variables
-sf_franks_mildred_clean <- sf_franks_mildred %>% select(Name = HNAME)
-sf_edb_reg_clean <- sf_edb_reg %>% select(Region = Regions)
+# Clean up shapefile of the open water regions with a more descriptive column header
+sf_ow_delta_c <- sf_ow_delta %>% rename(Region = HNAME)
 
-# Transform crs of Franks-Mildred and EDB region shapefiles to the crs of the HAB satellite data
+# Transform crs of open water regions shapefile to the crs of the HAB satellite data
 crs_hab_sat <- st_crs(df_hab_sat$strs_prx_obj[[1]])
-sf_franks_mildred_32611 <- st_transform(sf_franks_mildred_clean, crs = crs_hab_sat)
-sf_edb_reg_32611 <- st_transform(sf_edb_reg_clean, crs = crs_hab_sat)
+sf_ow_delta_32611 <- st_transform(sf_ow_delta_c, crs = crs_hab_sat)
 
-# Create a bounding box of the EDB regions shapefile which will be used to crop the satellite data
-  # and extend it by 5% on each side
-bbox_edb_reg <- st_bbox(sf_edb_reg_32611)
-bbox_edb_reg_xrange <- bbox_edb_reg$xmax - bbox_edb_reg$xmin
-bbox_edb_reg_yrange <- bbox_edb_reg$ymax - bbox_edb_reg$ymin
-bbox_edb_reg[1] <- bbox_edb_reg[1] - (bbox_edb_reg_xrange * 0.05)
-bbox_edb_reg[3] <- bbox_edb_reg[3] + (bbox_edb_reg_xrange * 0.05)
-bbox_edb_reg[2] <- bbox_edb_reg[2] - (bbox_edb_reg_yrange * 0.05)
-bbox_edb_reg[4] <- bbox_edb_reg[4] + (bbox_edb_reg_yrange * 0.05)
+# Create a bounding box of the open water regions shapefile which will be used to
+  # crop the satellite data. Add a 1 km buffer to slightly expand the bounding box
+  # to ensure no desired data is removed.
+bbox_ow_delta <- st_bbox(st_buffer(sf_ow_delta_32611, 1000))
 
 # Create a vector of dates to exclude from the analysis since the imagery doesn't cover the Delta region
 dates_rm <-
@@ -116,8 +109,9 @@ df_hab_sat_clean <- df_hab_sat %>%
   filter(!strs_date %in% dates_rm) %>%
   mutate(
     rast_obj_crop =
-      # Crop HAB satellite data to bounding box of the EDB regions to make it easier to work with
-      map(strs_prx_obj, ~st_crop(.x, bbox_edb_reg) %>%
+      # Crop HAB satellite data to bounding box of the open water regions
+        # shapefile to make it easier to work with
+      map(strs_prx_obj, ~st_crop(.x, bbox_ow_delta) %>%
         # rename attribute to be more descriptive
         setNames("pixel_val") %>%
         # Convert factor to numeric
@@ -129,92 +123,84 @@ df_hab_sat_clean <- df_hab_sat %>%
   ) %>%
   select(-strs_prx_obj)
 
-# Function to count number of pixels within each CI category and the percentage of valid pixels
-  # only for the pixels completely within the polygon (100% coverage fraction)
+# Function to count the number of pixels within each CI category for the pixels
+  # completely within the polygon (100% coverage fraction)
 count_ci_cat <- function(df) {
-  # Count pixels within each CI category
-  df_count <- df %>%
-    filter(coverage_fraction == 1) %>%
-    mutate(
-      ci_category = case_when(
-        value == 0 ~ "Non_detect",
-        value <= 41 ~ "Low",
-        value <= 99 ~ "Moderate",
-        value <= 183 ~ "High",
-        value <= 250 ~ "Very_high",
-        TRUE ~ "Invalid_or_missing"
-      )
-    ) %>%
-    count(ci_category) %>%
-    rename(pixel_count = n)
-
-  # Calculate total number of pixels counted
-  vec_pixel_sum <- sum(df_count$pixel_count)
-
-  df_f <- df_count %>%
-    # Restructure data frame into wide format
-    pivot_wider(names_from = ci_category, values_from = pixel_count) %>%
-    # Add total number of pixels counted and the percent of pixels with valid data (<= 250)
-    mutate(
-      pixel_sum = vec_pixel_sum,
-      perc_valid = (1 - Invalid_or_missing/pixel_sum) * 100
-    )
-
-  return(df_f)
-}
-
-# Function to replace NA counts with zeros for the CI categories
-repl_ci_count_na <- function(df) {
   df %>%
-    replace_na(
-      list(
-        Non_detect = 0,
-        Low = 0,
-        Moderate = 0,
-        High = 0,
-        Very_high = 0,
-        Invalid_or_missing = 0
-      )
-    )
+    filter(coverage_fraction == 1) %>%
+    left_join(EDBdata:::df_hab_sat_pixel_key, by = c("value" = "PixelValue")) %>%
+    count(CICategory, name = "pixel_count") %>%
+    pivot_wider(names_from = CICategory, values_from = pixel_count)
 }
 
-# Calculate counts of pixel values within each CI category for Franks Tract and Mildred Island
-hab_sat_fr_mil <- df_hab_sat_clean %>%
-  mutate(sf_fr_mil = list(sf_franks_mildred_32611)) %>%
+# Function to calculate the average Cyano Index of valid pixels completely
+  # within the polygon (100% coverage fraction)
+calc_avg_ci <- function(df) {
+  df %>%
+    filter(coverage_fraction == 1) %>%
+    left_join(EDBdata:::df_hab_sat_pixel_key, by = c("value" = "PixelValue")) %>%
+    summarize(AvgCI = mean(CyanoIndex, na.rm = TRUE)) %>%
+    pull(AvgCI)
+}
+
+# Finish preparing the HAB satellite data for the four open water regions in the Delta
+hab_sat_ow_delta <- df_hab_sat_clean %>%
+  mutate(sf_fr_mil_ccf_lib = list(sf_ow_delta_32611)) %>%
   # Extract pixels from within each polygon
-  mutate(sf_fr_mil = map2(sf_fr_mil, rast_obj_crop, ~mutate(.x, rast_extract = exact_extract(.y, .x)))) %>%
+  mutate(
+    sf_fr_mil_ccf_lib = map2(
+      sf_fr_mil_ccf_lib,
+      rast_obj_crop,
+      ~ mutate(.x, df_rast_extract = exact_extract(.y, .x))
+    )
+  ) %>%
   # Convert sf object to data frame
-  mutate(df_fr_mil = map(sf_fr_mil, st_drop_geometry)) %>%
-  select(Date = strs_date, df_fr_mil) %>%
-  unnest(cols = df_fr_mil) %>%
-  # Count number of pixels in each CI category, totals, and percent valid pixels
-  mutate(rast_extract = map(rast_extract, count_ci_cat)) %>%
-  # Simplify to a summary data frame
-  unnest(cols = rast_extract) %>%
+  mutate(df_fr_mil_ccf_lib = map(sf_fr_mil_ccf_lib, st_drop_geometry)) %>%
+  select(Date = strs_date, df_fr_mil_ccf_lib) %>%
+  unnest(cols = df_fr_mil_ccf_lib) %>%
+  # Count number of pixels in each CI category for each region and date
+  mutate(df_ci_count = map(df_rast_extract, count_ci_cat)) %>%
+  # Unnest CI category counts into data frame
+  unnest(cols = df_ci_count) %>%
+  # Replace NA values in the CI category counts with zeros
+  replace_na(
+    list(
+      NonDetect = 0,
+      Low = 0,
+      Moderate = 0,
+      High = 0,
+      VeryHigh = 0,
+      InvalidOrMissing = 0
+    )
+  ) %>%
+  # Calculate total number of pixels counted for each row and the percent of
+    # pixels with valid data (<= 250)
+  mutate(
+    PixelSum = rowSums(across(where(is.integer))),
+    PercValid = (1 - InvalidOrMissing/PixelSum) * 100
+  ) %>%
   # Only include days where there were greater than 25% valid pixels
-  filter(perc_valid > 25) %>%
+  filter(PercValid > 25) %>%
+  # Calculate the average Cyano Index value for each region and date
+  mutate(AvgCI = map_dbl(df_rast_extract, calc_avg_ci)) %>%
   # Reorder and select variables
   select(
     Date,
-    Name,
-    Non_detect,
+    Region,
+    AvgCI,
+    NonDetect,
     Low,
     Moderate,
     High,
-    Very_high,
-    Invalid_or_missing
-  ) %>%
-  # Replace NA counts with zeros
-  repl_ci_count_na()
+    VeryHigh,
+    InvalidOrMissing
+  )
 
-# Calculate counts of pixel values within each CI category for the EDB regions
-  # Not including this for now, but keeping it as a placeholder
-
-# Save final data set containing counts of pixel values within each CI category as csv file
+# Save final data set containing counts and averages of Cyano Index values as csv file
   # for easier diffing
-write_csv(hab_sat_fr_mil, here("data-raw/Final/hab_sat_fr_mil.csv"))
+write_csv(hab_sat_ow_delta, here("data-raw/Final/hab_sat_ow_delta.csv"))
 
-# Save final data sets containing counts of pixel values within each CI category as objects
+# Save final data set containing counts and averages of Cyano Index values as object
   # in the data package
-usethis::use_data(hab_sat_fr_mil, overwrite = TRUE)
+usethis::use_data(hab_sat_ow_delta, overwrite = TRUE)